refactor(P4-1/P4-3): ファイル名タイポ修正とデッドコード削除
P4-1（タイポ修正・命名規則統一）:
- iuser-repositroy.ts → iuser-repository.ts
- milestone.interfase.ts → imilestone-repository.ts
- domains/work-records/work-recoed.ts → work-record.ts
- import パスを全更新（本番・テスト計 ~22 参照）

P4-3（デッドコード削除）:
- domains/sync/IWbsSyncService.ts（本体コメントアウトの死んだポート、live参照0）
- applications/excel-sync/WbsSyncApplicationService.ts（44行コメントの旧実装、importer0）

据え置き（削除しない）:
- app/test/gantt/page.tsx: visual回帰テストのフィクスチャ（GanttVisualTestHelper が使用）
- api/notifications/analytics/route.ts: service-worker.js が使用中

本番型チェック・ユニットテスト 1292 件 green。
