+++ src/__tests__/domains/sync/WbsSyncService.test.ts
@@ -8,7 +8,7 @@
 // import { IExcelWbsRepository } from '@/applications/excel-sync/IExcelWbsRepository';
 // import { ISyncLogRepository } from '@/applications/excel-sync/ISyncLogRepository';
 // import { IPhaseRepository } from '@/applications/task/iphase-repository';
-// import { IUserRepository } from '@/applications/user/iuser-repositroy';
+// import { IUserRepository } from '@/applications/user/iuser-repository';
 // import { WbsAssignee } from '@/domains/wbs/wbs-assignee';
 // import { User } from '@/domains/user/user';
 // import { IWbsAssigneeRepository } from '@/applications/wbs/iwbs-assignee-repository';
