+++ src/__tests__/domains/task/task.test.ts
@@ -8,7 +8,7 @@ import { Period } from "@/domains/task/period";
 import { PeriodType } from "@/domains/task/value-object/period-type";
 import { ManHour } from "@/domains/task/man-hour";
 import { ManHourType } from "@/domains/task/value-object/man-hour-type";
-import { WorkRecord } from "@/domains/work-records/work-recoed";
+import { WorkRecord } from "@/domains/work-records/work-record";
 
 describe('Task', () => {
   const taskNo = TaskNo.reconstruct('D1-0001');
