--- src/domains/sync/IWbsSyncService.ts
@@ -1,6 +1,0 @@
-// import { SyncResult } from './ExcelWbs';;
-
-// export interface IWbsSyncService {
-//   // 洗い替え（全削除→全インポート）
-//   replaceAll(wbsId: number): Promise<SyncResult>;
-// }
