+++ src/domains/work-records/work-record.ts
