+++ src/infrastructures/geppo-import/user-mapping.service.ts
@@ -1,6 +1,6 @@
 import { injectable, inject } from 'inversify'
 import { SYMBOL } from '@/types/symbol'
-import type { IUserRepository } from '@/applications/user/iuser-repositroy'
+import type { IUserRepository } from '@/applications/user/iuser-repository'
 import type { IUserMappingService } from '@/applications/geppo-import/iuser-mapping-service'
 
 @injectable()
