+++ src/infrastructures/work-record/work-record-prisma.repository.ts
@@ -1,6 +1,6 @@
 import prisma from "@/lib/prisma/prisma";
 import { injectable} from 'inversify'
-import { WorkRecord } from '@/domains/work-records/work-recoed'
+import { WorkRecord } from '@/domains/work-records/work-record'
 import { IWorkRecordRepository } from '@/applications/work-record/repositories/iwork-record.repository'
 import type { WorkRecordDetail } from '@/types/work-record'
